Switch to using default min-layer-size
We were originally setting a min-layer-size of 0 when building
a SOCI index for an image. This was mainly due to the constraint
put forth by some registries that disallow pushing manifests
without any layers. To remain consistent with the CLI, we should
use the default min-layer-size (10 MiB). In the case where an
empty index is generated, we will log a warning indicating that
we are skipping pushing the empty index and exit gracefully.

Signed-off-by: Yasin Turan <turyasin@amazon.com>