Fix documented permission requirements
The documentation around the minimum permissions needed to deploy the
CloudFormation template drifted from the template itself and were
missing several resources/permissions. This adds those resources as well
as a note that the resource names might be truncated relative to the
permissions listed if the stack name is long.

Signed-off-by: Kern Walster <walster@amazon.com>